Make first pending test fail
Notice the hard-coded 3 in the test description. I am fine with it
because I do not expect this test to ever change, only deleted and
replaced with a more accurate/relevant one.

I take immutability to the extreme: immutable infrastructure, immutable
code, immutable tests. I strongly believe that composing simple
concepts that do not change makes for better codebases.