Add second failing test
This will force code that has a more useful behaviour. If you are
tempted to write "wrong" code that would pass the tests but isn't what
you we need, recognise that you are trolling yourself. Always remember
what you're trying to achieve, and take the smallest steps towards it.
The goal has been achieved when neither you or your pair can think of
any other tests to write, and the tests that you have are all passing.