Pass second test
In functions, unless a variable is explicitly defined as local, it will
become available outside the function (i.e. leaks into the global
scope).

The ${1:?error message} construct guards against undefined arguments.
It's not required in "private functions" - bash does not make the
distinction between private & public functions -, but it's better to
make it a habit at the expense of verbosity than to have to REPL in the
search of that mysterious bug. Having said that, push your guards
towards the edges and embrace confident code, it makes for a more
elegant codebase. As always, do what feels right, there is no one rule
that would apply to all scenarios.

To expand on ${1:?}, this will be familiar to many:

  if [ -z "$1" ]
  then
    echo "First agument requied"
    return 1
  else
    value="$1"
  fi

I favour this alternative since it achieves the same goal:

  value="${1:?First argument required}"