Return FizzBuzz for numbers divisible by 15
The only complex thing is where to place the line. The previous tests,
as well as this new test give me the confidence that I've put it in the
right place.

By this point you might be screaming "REFACTOR", but I want to have the
full picture, including all tests, before I take a step back.