Decompose fizzbuzz into smaller components
The initial refactoring, even if it didn't make the code any better, it
made me explore "Fizz" & "Buzz" composition alternatives. If you were to
learn one thing today, this is it:

TDD is your best companion in exploring the composability of simple concepts.

It's no accident that these 2 characteristics (simplicity &
composability) are at the core of all good codebases. We write tests so
that we are guided down this simple, "just enough" path. We write tests
so that we know when we have arrived.

I have removed all function argument guards and tests. Because these
functions are only called from within fizzbuzz, we have full control
over what gets passed into them. It's a balance and most often a
judgement call, but I personally prefer confident code that guards on
the edges of the system rather than everywhere. As always, do what feels
right.

I would also like to highlight the change of ordering in the code. I
find it pleasing when code runs left to right, top to bottom. Notice how
the functions are ordered outside-in: outermost layer at the top, innermost
layer at the bottom.